+++ lib/internal.h
@@ -82,7 +82,7 @@ void unshield_file_group_destroy(UnshieldFileGroup* self);
  */
 
 char *unshield_get_base_directory_name(Unshield *unshield);
-long int unshield_get_path_max(Unshield* unshield);
+size_t unshield_get_path_max(Unshield* unshield);
 FILE* unshield_fopen_for_reading(Unshield* unshield, int index, const char* suffix);
 long long unshield_fsize(FILE* file);
 bool unshield_read_common_header(uint8_t** buffer, CommonHeader* common);
